move map_unwrap_or to its own module
